Refactoring. Rename of bad package name
