+++ src/main/java/spark/webserver/serialization/DefaultSerializer.java
@@ -4,7 +4,7 @@
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
- *
+ *  
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -14,31 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spark.webserver.serializing;
+package spark.webserver.serialization;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
-
-import spark.utils.IOUtils;
+import java.io.UnsupportedEncodingException;
 
 /**
- * Input stream serializer.
+ * Serializer that writes the result of toString to output in UTF-8 encoding
  *
  * @author alex
  */
-class InputStreamSerializer extends Serializer {
+class DefaultSerializer extends Serializer {
 
     @Override
     public boolean canProcess(Object element) {
-        return element instanceof InputStream;
+        return true;
     }
 
     @Override
-    public void process(OutputStream outputStream, Object element)
-            throws IOException {
-        String content = IOUtils.toString((InputStream) element);
-        outputStream.write(content.getBytes("utf-8"));
+    public void process(OutputStream outputStream, Object element) throws IOException {
+        try {
+            outputStream.write(element.toString().getBytes("utf-8"));
+        } catch (UnsupportedEncodingException e) {
+            throw new IOException(e);
+        }
     }
 
 }
